Added nginx reverse proxy to enable access to SOCKS5 and HTTP ports over the network
This was previously not possible due to possibly the Mullvad app filtering IPs that was outside the docker host network subnet range. We resolve this by placing an nginx reverse proxy in front on the same host machine to proxy requests made from the network so that they "originate" from the host machine.