Fix Mullvad installation in Dockerfile
I'm not sure how the original Dockerfile worked at all, but the latest
Debian package requires systemd, because it installs and starts the
service in the postinst script.

See [1] (unrelated to Mullvad, but the same error) and [2].

This change extracts the binaries from the official package, and avoids
the systemd dependency.

[1]: https://github.com/microsoft/WSL/issues/8883

[2]: https://old.reddit.com/r/mullvadvpn/comments/16l3i5s/nonsystemd_support_for_docker_mullvad_client/